requirements.txt
# ============================================================
# itimecontrol-assistant — requirements.txt
# ============================================================

# --- Extracción de PDFs ---
pdfplumber>=0.10.3
PyMuPDF>=1.23.0          # fitz
pypdf>=3.17.0

# --- NLP y procesamiento de texto ---
nltk>=3.8.1
spacy>=3.7.2
langdetect>=1.0.9
ftfy>=6.1.3              # fix encoding issues

# --- Fine-Tuning ---
torch>=2.1.0
transformers>=4.38.0
datasets>=2.18.0
peft>=0.9.0              # LoRA / QLoRA
trl>=0.8.1               # SFTTrainer
bitsandbytes>=0.42.0     # cuantización 4-bit
accelerate>=0.27.0
sentencepiece>=0.2.0
protobuf>=4.25.0

# --- Embeddings y RAG ---
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4         # cambiar a faiss-gpu si hay GPU disponible
chromadb>=0.4.22

# --- API ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# --- Interfaz ---
gradio>=4.20.0

# --- Evaluación ---
rouge-score>=0.1.2
sacrebleu>=2.4.0

# --- Utilidades ---
python-dotenv>=1.0.1
pyyaml>=6.0.1
tqdm>=4.66.2
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.2

# --- Jupyter ---
jupyter>=1.0.0
ipykernel>=6.29.0
